+++ backend/file/entity/file.entity.ts
@@ -27,6 +27,7 @@ export class FileEntity {
 
   @Column()
   @IsNotEmpty()
+  @IsString()
   filePath!: string;
 
   @CreateDateColumn()